axfs

Module api

std::fs-like high-level filesystem manipulation operations.

Structs

DirBuilder

A builder used to create directories in various manners.

DirEntry

Entries returned by the ReadDir iterator.

File

An object providing access to an open file on the filesystem.

Metadata

Metadata information about a file.

OpenOptions

Options and flags which can be used to configure how a file is opened.

ReadDir

Iterator over the entries in a directory.

Functions

canonicalize

Returns the canonical, absolute form of a path with all intermediate components normalized.

create_dir

Creates a new, empty directory at the provided path.

create_dir_all

Recursively create a directory and all of its parent components if they are missing.

current_dir

Returns the current working directory as a String.

metadata

Given a path, query the file system to get information about a file, directory, etc.

read

Read the entire contents of a file into a bytes vector.

read_dir

Returns an iterator over the entries within a directory.

read_to_string

Read the entire contents of a file into a string.

remove_dir

Removes an empty directory.

remove_file

Removes a file from the filesystem.

rename

Rename a file or directory to a new name. Delete the original file if old already exists.

set_current_dir

Changes the current working directory to the specified path.

write

Write a slice as the entire contents of a file.

Type Aliases

FileType

A structure representing a type of file with accessors for each file type. It is returned by Metadata::file_type method.

Permissions

Representation of the various permissions on a file.