Crate axfs

ArceOS filesystem module.

It provides unified filesystem operations for various filesystems.

Cargo Features

fatfs: Use FAT as the main filesystem and mount it on /. This feature is enabled by default.

devfs: Mount [axfs_devfs::DeviceFileSystem] on /dev. This feature is enabled by default.

ramfs: Mount [axfs_ramfs::RamFileSystem] on /tmp. This feature is enabled by default.

myfs: Allow users to define their custom filesystems to override the default. In this case, MyFileSystemIf is required to be implemented to create and initialize other filesystems. This feature is disabled by by default, but it will override other filesystem selection features if both are enabled.

Modules

api

std::fs-like high-level filesystem manipulation operations.

fops

Low-level filesystem operations.

Functions

init_filesystems

Initializes filesystems by block devices.