axfs

Module fops

Low-level filesystem operations.

Structs

Directory

An opened directory object, with open permissions and a cursor for read_dir.

Diskmyfs

A disk device with a cursor.

File

An opened file object, with open permissions and a cursor.

OpenOptions

Options and flags which can be used to configure how a file is opened.

Traits

MyFileSystemIfmyfs

The interface to define custom filesystems in user apps.

Type Aliases

DirEntry

Alias of [axfs_vfs::VfsDirEntry].

FileAttr

Alias of [axfs_vfs::VfsNodeAttr].

FilePerm

Alias of [axfs_vfs::VfsNodePerm].

FileType

Alias of [axfs_vfs::VfsNodeType].